tem1 and tem2 are contains TPadj and Tpadjn files and are chi-square and color respectively

+++ pairwise/computeflow3.m
@@ -19,7 +19,7 @@ col=320;
     filename=regexprep(imname, '.png', '.segimage.mat');
     load(['./videoset/tem/',filename]);
     
-    nseg=max(max(segimage));
+    nseg=double(max(max(segimage)));
     filename=[namecell{iter+1},'.segimage.mat'];
     segimage1=importdata(['./videoset/tem/',filename]);        
     nseg1=max(max(segimage1));
